9497 - WIP need to call the right sequence to get the list of messages
